pyproject.toml
[project]
name = "quantumrings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "python-dotenv>=1.0.1",
    "qiskit>=1.3.2",
    "quantumringslib>=0.9.13",
]
